requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
requests==2.28.2
psycopg2-binary==2.9.6
shapely==2.0.1
pydantic==1.10.7
